+++ python/detectors/strawtubesDetector.py
@@ -1,30 +1,25 @@
 import ROOT
-import global_variables
 from BaseDetector import BaseDetector
 
-class timeDetector(BaseDetector):
+
+class strawtubesDetector(BaseDetector):
     def __init__(self, name, intree):
-        super().__init__(name, intree, 'std.vector')
+        super().__init__(name, intree, 'std.vector', 'Strawtubes')
 
     def digitize(self):
-        """Digitize timing detector MC hits.
+        """Digitize strawtube MC hits.
 
         The earliest hit per straw will be marked valid, all later ones invalid.
         """
         earliest_per_det_id = {}
-        for index, point in enumerate(self.intree.TimeDetPoint):
-            hit = ROOT.TimeDetHit(point, self.intree.t0)
+        for index, point in enumerate(self.intree.strawtubesPoint):
+            hit = ROOT.strawtubesHit(point, self.intree.t0)
             self.det.push_back(hit)
-            detector_id = hit.GetDetectorID()
             if hit.isValid():
+                detector_id = hit.GetDetectorID()
                 if detector_id in earliest_per_det_id:
-                    times = hit.GetMeasurements()
                     earliest = earliest_per_det_id[detector_id]
-                    reference_times = self.det[earliest].GetMeasurements()
-                    # this is not really correct, only first attempt
-                    # case that one measurement only is earlier not taken into account
-                    # SetTDC(Float_t val1, Float_t val2)
-                    if reference_times[0] > times[0] or reference_times[1] > times[1]:
+                    if self.det[earliest].GetTDC() > hit.GetTDC():
                         # second hit with smaller tdc
                         self.det[earliest].setInvalid()
                         earliest_per_det_id[detector_id] = index